requirements.txt
tensorflow==2.12.0
streamlit
numpy
Pillow
matplotlib